prevent rsync without $src set
If $src is not set here rsync is going to do a directory listing exiting with code 0, since only doing a directory listing clearly is not what we want to achieve we need to throw an exception. That way it is easier for continuous integration to recognize the failed deployment.