rsync recipe doesn't use identityFile
Even when set, the rsync command doesn't use the identityFile given from server configuration. Instead it will ask for password of the user.
Changed to use the identityFile. Passphrase would be asked anyway, as sshpass or any other similar program may not be available on the system. To use it the 'easy' way, creating a keyfile w/o passphrase would be a way, but security risk too.