Fix cloudflare purge command
Updated the cloudflare recipe to work again. not sure if it was broken due to an API change or what, but it was broken before.

The `api_key`, `email` and `domain` configs are equired. The `domain` config should be the `zoneId` which you can get by hitting this endpoint:

```bash
## Request Duplicate
curl "https://api.cloudflare.com/client/v4/zones" \
     -H 'X-Auth-Email: {YOUR EMAIL ADDRESS}' \
     -H 'X-Auth-Key: {YOUR API KEY} 
```